backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.40.0
websockets==16.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==3.2.0
joblib==1.3.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Utilities
uuid==1.30
yarl==1.22.0
orjson==3.9.10

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
